+++ src/commands/find.ts
@@ -1,7 +1,7 @@
 import * as vscode from "vscode";
 import { TextEditor } from "vscode";
 import { EmacsCommand } from ".";
-import { IEmacsCommandRunner, IMarkModeController } from "../emulator";
+import { IEmacsController } from "../emulator";
 import { MessageManager } from "../message";
 import { revealPrimaryActive } from "./helpers/reveal";
 
@@ -12,12 +12,8 @@ export interface SearchState {
 abstract class IsearchCommand extends EmacsCommand {
   protected searchState: SearchState;
 
-  public constructor(
-    afterExecute: () => void,
-    emacsController: IMarkModeController & IEmacsCommandRunner,
-    searchState: SearchState
-  ) {
-    super(afterExecute, emacsController);
+  public constructor(emacsController: IEmacsController, searchState: SearchState) {
+    super(emacsController);
 
     this.searchState = searchState;
   }
@@ -29,7 +25,14 @@ export class IsearchForward extends IsearchCommand {
   public execute(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined): Thenable<void> {
     this.searchState.startSelections = textEditor.selections;
     return vscode.commands
-      .executeCommand("actions.find")
+      .executeCommand("editor.actions.findWithArgs", {
+        searchString: undefined,
+        replaceString: undefined,
+        isRegex: false,
+        matchWholeWord: false,
+        isCaseSensitive: false,
+        preserveCase: false,
+      })
       .then(() => vscode.commands.executeCommand<void>("editor.action.nextMatchFindAction"));
   }
 }
@@ -40,8 +43,81 @@ export class IsearchBackward extends IsearchCommand {
   public execute(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined): Thenable<void> {
     this.searchState.startSelections = textEditor.selections;
     return vscode.commands
-      .executeCommand("actions.find")
+      .executeCommand("editor.actions.findWithArgs", {
+        searchString: undefined,
+        replaceString: undefined,
+        isRegex: false,
+        matchWholeWord: false,
+        isCaseSensitive: false,
+        preserveCase: false,
+      })
       .then(() => vscode.commands.executeCommand<void>("editor.action.previousMatchFindAction"));
+  }
+}
+
+export class IsearchForwardRegexp extends IsearchCommand {
+  public readonly id = "isearchForwardRegexp";
+
+  public execute(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined): Thenable<void> {
+    this.searchState.startSelections = textEditor.selections;
+    return vscode.commands
+      .executeCommand("editor.actions.findWithArgs", {
+        searchString: undefined,
+        replaceString: undefined,
+        isRegex: true,
+        matchWholeWord: false,
+        isCaseSensitive: false,
+        preserveCase: false,
+      })
+      .then(() => vscode.commands.executeCommand<void>("editor.action.nextMatchFindAction"));
+  }
+}
+
+export class IsearchBackwardRegexp extends IsearchCommand {
+  public readonly id = "isearchBackwardRegexp";
+
+  public execute(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined): Thenable<void> {
+    this.searchState.startSelections = textEditor.selections;
+    return vscode.commands
+      .executeCommand("editor.actions.findWithArgs", {
+        searchString: undefined,
+        replaceString: undefined,
+        isRegex: true,
+        matchWholeWord: false,
+        isCaseSensitive: false,
+        preserveCase: false,
+      })
+      .then(() => vscode.commands.executeCommand<void>("editor.action.previousMatchFindAction"));
+  }
+}
+
+export class QueryReplace extends IsearchCommand {
+  public readonly id = "queryReplace";
+
+  public execute(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined): Thenable<void> {
+    this.searchState.startSelections = textEditor.selections;
+    // I could not find a way to open the find widget with `editor.actions.findWithArgs`
+    // revealing the replace input and restoring the both query and replace strings.
+    // So `editor.action.startFindReplaceAction` is used here.
+    return vscode.commands.executeCommand("editor.action.startFindReplaceAction");
+  }
+}
+
+export class QueryReplaceRegexp extends IsearchCommand {
+  public readonly id = "queryReplaceRegexp";
+
+  public execute(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined): Thenable<void> {
+    this.searchState.startSelections = textEditor.selections;
+    // Like `queryReplace` command, I could not find a way to open the find widget with the desired state.
+    // In this command, setting `isRegex` is the priority and I gave up restoring the replace string by setting ´replaceString=undefined`.
+    return vscode.commands.executeCommand("editor.actions.findWithArgs", {
+      searchString: undefined,
+      replaceString: "",
+      isRegex: true,
+      matchWholeWord: false,
+      isCaseSensitive: false,
+      preserveCase: false,
+    });
   }
 }