+++ src/commands/index.ts
@@ -1,5 +1,5 @@
 import { TextEditor } from "vscode";
-import { IEmacsCommandRunner, IMarkModeController } from "../emulator";
+import { IEmacsController } from "../emulator";
 
 export function createParallel<T>(concurrency: number, promiseFactory: () => Thenable<T>): Thenable<T[]> {
   return Promise.all(Array.from({ length: concurrency }, promiseFactory));
@@ -8,21 +8,18 @@ export function createParallel<T>(concurrency: number, promiseFactory: () => The
 export abstract class EmacsCommand {
   public abstract readonly id: string;
 
-  protected emacsController: IMarkModeController & IEmacsCommandRunner;
-  private afterExecute: () => void | Promise<unknown>;
+  protected emacsController: IEmacsController;
 
-  public constructor(afterExecute: () => void, markModeController: IMarkModeController & IEmacsCommandRunner) {
-    this.afterExecute = afterExecute;
+  public constructor(markModeController: IEmacsController) {
     this.emacsController = markModeController;
   }
 
-  public run(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined): Thenable<unknown> {
-    const ret = this.execute(textEditor, isInMarkMode, prefixArgument);
-    if (ret != null) {
-      return ret.then(this.afterExecute);
-    } else {
-      return Promise.resolve(this.afterExecute());
-    }
+  public run(
+    textEditor: TextEditor,
+    isInMarkMode: boolean,
+    prefixArgument: number | undefined
+  ): Thenable<unknown> | void {
+    return this.execute(textEditor, isInMarkMode, prefixArgument);
   }
 
   public abstract execute(