+++ src/commands/rectangle.ts
@@ -1,7 +1,7 @@
 import * as vscode from "vscode";
 import { TextEditor } from "vscode";
 import { EmacsCommand, IEmacsCommandInterrupted } from ".";
-import { IEmacsCommandRunner, IMarkModeController } from "../emulator";
+import { IEmacsController } from "../emulator";
 import { getNonEmptySelections, makeSelectionsEmpty } from "./helpers/selection";
 import { convertSelectionToRectSelections } from "../rectangle";
 import { revealPrimaryActive } from "./helpers/reveal";
@@ -50,12 +50,8 @@ export interface RectangleState {
 export abstract class RectangleKillYankCommand extends EmacsCommand {
   protected rectangleState: RectangleState;
 
-  public constructor(
-    afterExecute: () => void,
-    emacsController: IMarkModeController & IEmacsCommandRunner,
-    rectangleState: RectangleState
-  ) {
-    super(afterExecute, emacsController);
+  public constructor(emacsController: IEmacsController, rectangleState: RectangleState) {
+    super(emacsController);
 
     this.rectangleState = rectangleState;
   }
@@ -256,12 +252,8 @@ export class StringRectangle extends EmacsCommand {
 
   private minibuffer: Minibuffer;
 
-  constructor(
-    afterExecute: () => void,
-    markModeController: IMarkModeController & IEmacsCommandRunner,
-    minibuffer: Minibuffer
-  ) {
-    super(afterExecute, markModeController);
+  constructor(markModeController: IEmacsController, minibuffer: Minibuffer) {
+    super(markModeController);
     this.minibuffer = minibuffer;
   }
 
@@ -295,12 +287,8 @@ export class ReplaceKillRingToRectangle extends EmacsCommand {
   public readonly id = "replaceKillRingToRectangle";
   private killring: KillRing | null;
 
-  public constructor(
-    afterExecute: () => void,
-    emacsController: IMarkModeController & IEmacsCommandRunner,
-    killring: KillRing | null
-  ) {
-    super(afterExecute, emacsController);
+  public constructor(emacsController: IEmacsController, killring: KillRing | null) {
+    super(emacsController);
     this.killring = killring;
   }